feat: add unit to product & change quantity to be double instead of int

+++ src/main/java/fr/ingeniance/katas/supermarketpricer/models/IOffer.java
@@ -3,5 +3,5 @@ package fr.ingeniance.katas.supermarketpricer.models;
 import java.math.BigDecimal;
 
 public interface IOffer {
-	public BigDecimal calculatePrice(Product product, int quantity);
+	public BigDecimal calculatePrice(Product product, double quantity);
 }
